added GetSocietiesForUser

+++ internal/rpc/contract.go
@@ -8,7 +8,7 @@ type DbRepo interface {
 	CreateGroup(socData *model.SocietyData) (int, error)
 	GetAccessLevel() (*[]model.AccessLevel, error)
 	GetPermissions() (*[]model.GetPermissions, error)
-	GetSocietyWithOffset(data *model.WithOffsetData) (*[]model.SocietyWithOffsetData, int, error)
+	GetSocietyWithOffset(data *model.WithOffsetData) (*[]model.SocietyWithOffsetData, int64, error)
 	GetSocietyInfo(id int64) (*model.SocietyInfo, error)
 	SubscribeToSociety(id int64, uuid string) (bool, error)
 	UnsubscribeFromSociety(id int64, uuid string) (bool, error)
